Fix P/Invoke signatures for archive_read_data and archive_write_data
Both functions return la_ssize_t (pointer-sized signed int) and take
size_t parameters. Using int truncates on 64-bit systems.

Changed signatures from:
  int archive_write_data(..., int length) -> int
  int archive_read_data(..., int size) -> int

To:
  nint archive_write_data(..., nint length) -> nint
  nint archive_read_data(..., nint size) -> nint

Updated call sites to cast nint results back to int for local arithmetic.

+++ LibArchive.Net/LibArchiveReader.cs
@@ -280,7 +280,7 @@ public partial class LibArchiveReader : SafeHandleZeroOrMinusOneIsInvalid
             if (_eof)
                 return 0;
 
-            int result = archive_read_data(archiveHandle, ref buffer[offset], count);
+            nint result = archive_read_data(archiveHandle, ref buffer[offset], count);
             if (result < 0)
             {
                 // Negative return indicates error (e.g., wrong password, encryption not supported)
@@ -292,7 +292,7 @@ public partial class LibArchiveReader : SafeHandleZeroOrMinusOneIsInvalid
             if (result == 0)
                 _eof = true;
 
-            return result;
+            return (int)result;
         }
 
         /// <summary>
@@ -377,7 +377,7 @@ public partial class LibArchiveReader : SafeHandleZeroOrMinusOneIsInvalid
     private static partial int archive_read_open_filenames(IntPtr a, IntPtr filename, int blocksize);
 
     [LibraryImport("archive")]
-    private static partial int archive_read_data(IntPtr a, ref byte buff, int size);
+    private static partial nint archive_read_data(IntPtr a, ref byte buff, nint size);
 
     [LibraryImport("archive")]
     private static partial int archive_read_next_header(IntPtr a, out IntPtr entry);
@@ -416,7 +416,7 @@ public partial class LibArchiveReader : SafeHandleZeroOrMinusOneIsInvalid
     private static extern int archive_read_open_filenames(IntPtr a, IntPtr filename, int blocksize);
 
     [DllImport("archive")]
-    private static extern int archive_read_data(IntPtr a, ref byte buff, int size);
+    private static extern nint archive_read_data(IntPtr a, ref byte buff, nint size);
 
     [DllImport("archive")]
     private static extern int archive_read_next_header(IntPtr a, out IntPtr entry);
